fix: fix incorrectly published version for angular 15
