+++ .gitignore
@@ -1,12 +1,6 @@
-/.angular/cache
-/node_modules
-/dist
-/dist-demo
-/documentation
-/coverage
-/.idea
-
-*.log
-*.tgz
-
-projects/ngx-strength-meter/README.md
+.idea
+node_modules
+docs/build/
+dist
+.angular
+coverage
