--- .npmignore
@@ -1,8 +1,0 @@
-/tests
-/spec.bundle.js
-/karma.conf.js
-.travis.yml
-/.idea
-/coverage
-*.log
-*.tgz
