+++ .npmrc
@@ -1,0 +1,1 @@
+@eisberg-labs:registry=https://registry.npmjs.org/
