+++ CHANGELOG.md
@@ -1,0 +1,11 @@
+# Change Log
+
+All notable changes to this project will be documented in this file.
+See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+# 15.0.0 (2022-12-14)
+
+
+### Features
+
+* update to Angular 15 ([4e5f954](https://github.com/eisberg-labs/ngx-strength-meter/commit/4e5f9542f50aff5a40776c51f414996bafa4a821))
