+++ ngx-strength-meter/ng-package.json
@@ -1,6 +1,6 @@
 {
   "$schema": "../../node_modules/ng-packagr/ng-package.schema.json",
-  "dest": "../../dist/ngx-strength-meter",
+  "dest": "../dist/ngx-strength-meter",
   "lib": {
     "entryFile": "src/public-api.ts"
   },
