+++ ngx-strength-meter/package.json
@@ -1,18 +1,18 @@
 {
-  "name": "ngx-strength-meter",
+  "name": "@eisberg-labs/ngx-strength-meter",
   "private": false,
-  "version": "13.1.2",
+  "version": "15.0.1",
   "description": "Angular widget for password strength estimation using zxcvbn",
   "dependencies": {
     "tslib": "^2.3.0",
-    "zxcvbn": "4.4.2"
+    "zxcvbn": "^4.4.2"
   },
   "peerDependencies": {
-    "@angular/core": "^13.2.2",
-    "@angular/common": "^13.2.2"
+    "@angular/common": "^15.2.0",
+    "@angular/core": "^15.2.0"
   },
   "author": {
-    "name": "Eisberg Labs",
+    "name": "Ana Bujan Eisberg Labs",
     "email": "ana@eisberg-labs.com",
     "url": "https://www.eisberg-labs.com"
   },
@@ -23,14 +23,15 @@
   "bugs": {
     "url": "https://github.com/eisberg-labs/ngx-strength-meter/issues"
   },
-  "homepage": "https://github.com/eisberg-labs/ngx-strength-meter",
+  "homepage": "https://ngx-strength-meter.amarjanica.com",
   "keywords": [
     "angular",
     "javascript",
     "typescript",
     "zxcvbn",
     "password",
-    "strength"
+    "strength",
+    "password-strength"
   ],
   "license": "MIT"
 }
