+++ ngx-strength-meter/src/lib/strength-meter.component.spec.ts
@@ -1,5 +1,5 @@
-import {ComponentFixture, TestBed} from '@angular/core/testing';
-import {StrengthMeterComponent} from './strength-meter.component';
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { StrengthMeterComponent } from './strength-meter.component';
 
 describe('StrengthMeterComponent', () => {
   let component: StrengthMeterComponent;
@@ -8,7 +8,7 @@ describe('StrengthMeterComponent', () => {
   beforeEach(() => {
     TestBed.configureTestingModule({
       imports: [],
-      declarations: [StrengthMeterComponent]
+      declarations: [StrengthMeterComponent],
     }).compileComponents();
 
     fixture = TestBed.createComponent(StrengthMeterComponent);
