+++ ngx-strength-meter/src/lib/strength-meter.module.ts
@@ -1,0 +1,11 @@
+import { NgModule } from '@angular/core';
+import { StrengthMeterComponent } from './strength-meter.component';
+
+const COMPONENTS = [StrengthMeterComponent];
+
+@NgModule({
+  imports: [],
+  declarations: COMPONENTS,
+  exports: COMPONENTS,
+})
+export class StrengthMeterModule {}
