+++ ngx-strength-meter/src/public-api.ts
@@ -1,0 +1,2 @@
+export * from './lib/strength-meter.module';
+export * from './lib/strength-meter.component';
