+++ ngx-strength-meter/tsconfig.base.json
@@ -1,0 +1,9 @@
+{
+  "extends": "../tsconfig.json",
+  "compilerOptions": {
+    "noImplicitAny": false
+  },
+  "files": [
+    "./typings.d.ts"
+  ]
+}
