+++ ngx-strength-meter/tsconfig.spec.json
@@ -1,7 +1,8 @@
+/* To learn more about this file see: https://angular.io/config/tsconfig. */
 {
-  "extends": "../../tsconfig.json",
+  "extends": "./tsconfig.base.json",
   "compilerOptions": {
-    "outDir": "../../out-tsc/spec",
+    "outDir": "../out-tsc/spec",
     "types": [
       "jasmine",
       "node"
