+++ ngx-strength-meter/typings.d.ts
@@ -1,0 +1,3 @@
+declare var zxcvbn: any;
+declare module 'zxcvbn';
+
