--- projects/ngx-strength-meter/.release-it.json
@@ -1,11 +1,0 @@
-{
-  "git": {
-    "commitMessage": "chore: release v${version}"
-  },
-  "github": {
-    "release": true
-  },
-  "npm": {
-    "publish": false
-  }
-}
