--- projects/ngx-strength-meter/src/public-api.ts
@@ -1,9 +1,0 @@
-/**
- * Copyright Eisberg Labs
- * MIT license
- * https://github.com/eisberg-labs/ngx-strength-meter
- *
- * Entry point for all public APIs of the package.
- */
-export * from './lib/strength-meter.module';
-export * from './lib/strength-meter.component';
