--- projects/ngx-strength-meter/tsconfig.lib.prod.json
@@ -1,6 +1,0 @@
-{
-  "extends": "./tsconfig.lib.json",
-  "angularCompilerOptions": {
-    "enableIvy": false
-  }
-}
