--- projects/ngx-strength-meter/typings.d.ts
@@ -1,3 +1,0 @@
-declare var zxcvbn: any;
-declare module 'zxcvbn/dist/zxcvbn.js';
-
