+++ renovate.json
@@ -1,0 +1,9 @@
+{
+  "$schema": "https://docs.renovatebot.com/renovate-schema.json",
+  "extends": ["config:base"],
+  "packageRules": [ {
+    "packagePatterns": ["*"],
+    "excludePackagePatterns": ["^@angular.+", "zxcvbn"],
+    "enabled": false
+  }]
+}
