--- src/app/app.component.html
@@ -1,24 +1,0 @@
-<div class="flex-container">
-  <div class="header flex-item">
-    <h2 style='text-align: center;'>
-      NGX STRENGTH METER
-    </h2>
-
-    <p style="text-align: center; margin-top: 20px;">
-      Type any password to see strength meter in action.
-    </p>
-  </div>
-  <div class="flex-item" style="max-width:300px;margin:0 auto">
-    <mat-card>
-      <mat-card-content>
-        <form [formGroup]="form" name="form">
-          <mat-form-field style="width:100%">
-            <input name="password" matInput placeholder="Input password" type="password" formControlName="password">
-          </mat-form-field>
-          <strength-meter [value]="form.value.password" (measure)="onStrengthChange($event)"></strength-meter>
-        </form>
-      </mat-card-content>
-    </mat-card>
-  </div>
-</div>
-
