--- src/app/app.component.ts
@@ -1,24 +1,0 @@
-import {Component, OnInit} from '@angular/core';
-import {FormBuilder, FormGroup, Validators} from '@angular/forms';
-
-@Component({
-  selector: 'app-root',
-  templateUrl: './app.component.html',
-  styleUrls: ['./app.component.css']
-})
-export class AppComponent implements OnInit {
-  form: FormGroup;
-
-  constructor(private fb: FormBuilder) {
-  }
-
-  ngOnInit(): void {
-    this.form = this.fb.group({
-      password: ['', Validators.required],
-    });
-  }
-
-  onStrengthChange(score) {
-    console.log('new score', score);
-  }
-}
