--- src/tsconfig.app.json
@@ -1,17 +1,0 @@
-{
-  "extends": "../tsconfig.json",
-  "compilerOptions": {
-    "outDir": "../out-tsc/app",
-    "module": "es2015",
-    "allowSyntheticDefaultImports": true,
-    "types": []
-  },
-  "exclude": [
-    "src/test.ts",
-    "**/*.spec.ts"
-  ],
-  "files": [
-    "polyfills.ts",
-    "main.ts"
-  ]
-}
